ccl: move pkg/ccl/partitionccl to pkg/sql/partitioning
Informs #164337
Release note: None
