+++ pkg/BUILD.bazel
@@ -72,7 +72,6 @@ ALL_TESTS = [
     "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/oidcccl:oidcccl_test",
-    "//pkg/ccl/partitionccl:partitionccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
@@ -611,6 +610,7 @@ ALL_TESTS = [
     "//pkg/sql/parser:parser_disallowed_imports_test",
     "//pkg/sql/parser:parser_test",
     "//pkg/sql/parserutils:parserutils_disallowed_imports_test",
+    "//pkg/sql/partitioning:partitioning_test",
     "//pkg/sql/pgcrypto/pgcryptocipher:pgcryptocipher_test",
     "//pkg/sql/pgcrypto:pgcrypto_test",
     "//pkg/sql/pgrepl/pgreplparser:pgreplparser_test",
@@ -1056,8 +1056,6 @@ GO_TARGETS = [
     "//pkg/ccl/multitenantccl:multitenantccl",
     "//pkg/ccl/oidcccl:oidcccl",
     "//pkg/ccl/oidcccl:oidcccl_test",
-    "//pkg/ccl/partitionccl:partitionccl",
-    "//pkg/ccl/partitionccl:partitionccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl",
@@ -2396,6 +2394,8 @@ GO_TARGETS = [
     "//pkg/sql/parser:parser",
     "//pkg/sql/parser:parser_test",
     "//pkg/sql/parserutils:parserutils",
+    "//pkg/sql/partitioning:partitioning",
+    "//pkg/sql/partitioning:partitioning_test",
     "//pkg/sql/pgcrypto/pgcryptocipher:pgcryptocipher",
     "//pkg/sql/pgcrypto/pgcryptocipher:pgcryptocipher_test",
     "//pkg/sql/pgcrypto:pgcrypto",
