+++ pkg/backup/BUILD.bazel
@@ -235,7 +235,6 @@ go_test(
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/multiregionccl/multiregionccltestutils",
         "//pkg/ccl/multitenantccl",
-        "//pkg/ccl/partitionccl",
         "//pkg/ccl/utilccl",
         "//pkg/cloud",
         "//pkg/cloud/amazon",
