+++ pkg/ccl/multiregionccl/BUILD.bazel
@@ -48,7 +48,6 @@ go_test(
         "//pkg/base",
         "//pkg/ccl",
         "//pkg/ccl/multiregionccl/multiregionccltestutils",
-        "//pkg/ccl/partitionccl",
         "//pkg/ccl/testutilsccl",
         "//pkg/clusterversion",
         "//pkg/jobs",
