+++ pkg/ccl/multiregionccl/show_test.go
@@ -8,8 +8,6 @@ package multiregionccl_test
 import (
 	"testing"
 
-	// Blank import partitionccl to install CreatePartitioning hook.
-	_ "github.com/cockroachdb/cockroach/pkg/ccl/partitionccl"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqltestutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
