+++ pkg/ccl/spanconfigccl/spanconfiglimiterccl/BUILD.bazel
@@ -11,7 +11,6 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/ccl/partitionccl",
         "//pkg/config",
         "//pkg/config/zonepb",
         "//pkg/roachpb",
