+++ pkg/ccl/spanconfigccl/spanconfigreconcilerccl/BUILD.bazel
@@ -10,7 +10,6 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/ccl/partitionccl",
         "//pkg/jobs",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
