+++ pkg/ccl/spanconfigccl/spanconfigsplitterccl/BUILD.bazel
@@ -10,7 +10,6 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/ccl/partitionccl",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
