+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/BUILD.bazel
@@ -12,7 +12,6 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/ccl/partitionccl",
         "//pkg/clusterversion",
         "//pkg/config/zonepb",
         "//pkg/jobs/jobspb",
