--- pkg/sql/logictest/testdata/logic_test/partitioning
@@ -1,17 +1,0 @@
-# enterprise-configs are blocked from running this file because they run with
-# a CCL binary, so the expected failures from using a non-CCL binary don't occur.
-# LogicTest: !enterprise-configs
-
-statement error pgcode XXC01 creating or manipulating partitions requires a CCL binary
-CREATE TABLE t (a INT, b INT, c INT, PRIMARY KEY (a, b)) PARTITION BY LIST (a) (
-    PARTITION p1 VALUES IN (1),
-    PARTITION p2 VALUES IN (2)
-)
-
-statement error pgcode XXC01 creating or manipulating partitions requires a CCL binary
-CREATE TABLE t (
-    a INT PRIMARY KEY, b INT,
-    INDEX (b) PARTITION BY LIST (b) (
-        PARTITION p1 VALUES IN (1)
-    )
-)
