+++ pkg/sql/logictest/tests/fakedist-vec-off/generated_test.go
@@ -1417,13 +1417,6 @@ func TestLogic_partial_txn_commit(
 	runLogicTest(t, "partial_txn_commit")
 }
 
-func TestLogic_partitioning(
-	t *testing.T,
-) {
-	defer leaktest.AfterTest(t)()
-	runLogicTest(t, "partitioning")
-}
-
 func TestLogic_pg_builtins(
 	t *testing.T,
 ) {
