+++ pkg/sql/logictest/tests/local-mixed-25.4/generated_test.go
@@ -1424,13 +1424,6 @@ func TestLogic_partial_txn_commit(
 	runLogicTest(t, "partial_txn_commit")
 }
 
-func TestLogic_partitioning(
-	t *testing.T,
-) {
-	defer leaktest.AfterTest(t)()
-	runLogicTest(t, "partitioning")
-}
-
 func TestLogic_pg_builtins(
 	t *testing.T,
 ) {
