+++ pkg/sql/partitioning/alter_primary_key_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package partitionccl
+package partitioning_test
 
 import (
 	"testing"
