+++ pkg/sql/partitioning/drop_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package partitionccl
+package partitioning_test
 
 import (
 	"context"
@@ -216,7 +216,7 @@ SELECT job_id
 	              PARTITION c VALUES IN ('c')
 	  )`)
 		tdb.Exec(t, `CREATE INDEX idx ON t (e)`)
-		tdb.Exec(t, `ALTER PARTITION a OF TABLE t CONFIGURE ZONE USING range_min_bytes = 123456, 
+		tdb.Exec(t, `ALTER PARTITION a OF TABLE t CONFIGURE ZONE USING range_min_bytes = 123456,
 range_max_bytes = 654321000`)
 		tdb.Exec(t, `ALTER INDEX t@idx CONFIGURE ZONE USING gc.ttlseconds = 1`)
 		tdb.Exec(t, `DROP INDEX t@idx`)
@@ -272,7 +272,7 @@ range_max_bytes = 654321000`)
 		)
 		tdb.Exec(t, `ALTER PARTITION ai OF INDEX t@idx CONFIGURE ZONE USING range_min_bytes = 123456,
 range_max_bytes = 654321000`)
-		tdb.Exec(t, `ALTER PARTITION a OF TABLE t CONFIGURE ZONE USING range_min_bytes = 123456, 
+		tdb.Exec(t, `ALTER PARTITION a OF TABLE t CONFIGURE ZONE USING range_min_bytes = 123456,
 range_max_bytes = 654321000`)
 		tdb.Exec(t, `ALTER INDEX t@idx CONFIGURE ZONE USING gc.ttlseconds = 1`)
 		tdb.Exec(t, `DROP INDEX t@idx`)
