+++ pkg/sql/schemachanger/scdeps/sctestdeps/BUILD.bazel
@@ -24,7 +24,6 @@ go_library(
         "//pkg/settings/cluster",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/catalogkeys",
-        "//pkg/sql/catalog/catpb",
         "//pkg/sql/catalog/dbdesc",
         "//pkg/sql/catalog/descbuilder",
         "//pkg/sql/catalog/descidgen",
