+++ pkg/sql/schemachanger/scdeps/sctestdeps/test_deps.go
@@ -23,7 +23,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkeys"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descbuilder"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/funcdesc"
@@ -250,27 +249,6 @@ func (s *TestState) HasGlobalPrivilegeOrRoleOption(
 func (s *TestState) CheckRoleExists(ctx context.Context, role username.SQLUsername) error {
 	s.LogSideEffectf("checking role/user %q exists", role)
 	return nil
-}
-
-// IndexPartitioningCCLCallback implements the scbuild.Dependencies interface.
-func (s *TestState) IndexPartitioningCCLCallback() scbuild.CreatePartitioningCCLCallback {
-	if ccl := scdeps.CreatePartitioningCCL; ccl != nil {
-		return ccl
-	}
-	return func(
-		ctx context.Context,
-		st *cluster.Settings,
-		evalCtx *eval.Context,
-		columnLookupFn func(tree.Name) (catalog.Column, error),
-		oldNumImplicitColumns int,
-		oldKeyColumnNames []string,
-		partBy *tree.PartitionBy,
-		allowedNewColumnNames []tree.Name,
-		allowImplicitPartitioning bool,
-	) (newImplicitCols []catalog.Column, newPartitioning catpb.PartitioningDescriptor, err error) {
-		newPartitioning.NumColumns = uint32(len(partBy.Fields))
-		return nil, newPartitioning, nil
-	}
 }
 
 var _ scbuild.CatalogReader = (*TestState)(nil)
